+++ pages/all-lots.html
@@ -67,7 +67,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Доски и лыжи</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">2014 Rossignol District Snowboard</a></h3>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">2014 Rossignol District Snowboard</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
                   <span class="lot__amount">Стартовая цена</span>
@@ -85,7 +85,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Доски и лыжи</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">DC Ply Mens 2016/2017 Snowboard</a></h3>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">DC Ply Mens 2016/2017 Snowboard</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
                   <span class="lot__amount">12 ставок</span>
@@ -103,7 +103,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Крепления</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Крепления Union Contact Pro 2015 года размер
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Крепления Union Contact Pro 2015 года размер
                 L/XL</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
@@ -122,7 +122,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Ботинки</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Ботинки для сноуборда DC Mutiny Charocal</a>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Ботинки для сноуборда DC Mutiny Charocal</a>
               </h3>
               <div class="lot__state">
                 <div class="lot__rate">
@@ -141,7 +141,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Одежда</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Куртка для сноуборда DC Mutiny Charocal</a></h3>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Куртка для сноуборда DC Mutiny Charocal</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
                   <span class="lot__amount">12 ставок</span>
@@ -159,7 +159,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Разное</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Маска Oakley Canopy</a></h3>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Маска Oakley Canopy</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
                   <span class="lot__amount">Стартовая цена</span>
@@ -177,7 +177,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Ботинки</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Ботинки для сноуборда DC Mutiny Charocal</a>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Ботинки для сноуборда DC Mutiny Charocal</a>
               </h3>
               <div class="lot__state">
                 <div class="lot__rate">
@@ -196,7 +196,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Одежда</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Куртка для сноуборда DC Mutiny Charocal</a></h3>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Куртка для сноуборда DC Mutiny Charocal</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
                   <span class="lot__amount">12 ставок</span>
@@ -214,7 +214,7 @@
             </div>
             <div class="lot__info">
               <span class="lot__category">Разное</span>
-              <h3 class="lot__title"><a class="text-link" href="lot.html">Маска Oakley Canopy</a></h3>
+              <h3 class="lot__title"><a class="text-link" href="lot.php">Маска Oakley Canopy</a></h3>
               <div class="lot__state">
                 <div class="lot__rate">
                   <span class="lot__amount">Стартовая цена</span>
