+++ pages/index.html
@@ -66,7 +66,7 @@
           </div>
           <div class="lot__info">
             <span class="lot__category">Доски и лыжи</span>
-            <h3 class="lot__title"><a class="text-link" href="lot.html">2014 Rossignol District Snowboard</a></h3>
+            <h3 class="lot__title"><a class="text-link" href="lot.php">2014 Rossignol District Snowboard</a></h3>
             <div class="lot__state">
               <div class="lot__rate">
                 <span class="lot__amount">Стартовая цена</span>
@@ -84,7 +84,7 @@
           </div>
           <div class="lot__info">
             <span class="lot__category">Доски и лыжи</span>
-            <h3 class="lot__title"><a class="text-link" href="lot.html">DC Ply Mens 2016/2017 Snowboard</a></h3>
+            <h3 class="lot__title"><a class="text-link" href="lot.php">DC Ply Mens 2016/2017 Snowboard</a></h3>
             <div class="lot__state">
               <div class="lot__rate">
                 <span class="lot__amount">12 ставок</span>
@@ -102,7 +102,7 @@
           </div>
           <div class="lot__info">
             <span class="lot__category">Крепления</span>
-            <h3 class="lot__title"><a class="text-link" href="lot.html">Крепления Union Contact Pro 2015 года размер L/XL</a></h3>
+            <h3 class="lot__title"><a class="text-link" href="lot.php">Крепления Union Contact Pro 2015 года размер L/XL</a></h3>
             <div class="lot__state">
               <div class="lot__rate">
                 <span class="lot__amount">7 ставок</span>
@@ -120,7 +120,7 @@
           </div>
           <div class="lot__info">
             <span class="lot__category">Ботинки</span>
-            <h3 class="lot__title"><a class="text-link" href="lot.html">Ботинки для сноуборда DC Mutiny Charocal</a></h3>
+            <h3 class="lot__title"><a class="text-link" href="lot.php">Ботинки для сноуборда DC Mutiny Charocal</a></h3>
             <div class="lot__state">
               <div class="lot__rate">
                 <span class="lot__amount">12 ставок</span>
@@ -138,7 +138,7 @@
           </div>
           <div class="lot__info">
             <span class="lot__category">Одежда</span>
-            <h3 class="lot__title"><a class="text-link" href="lot.html">Куртка для сноуборда DC Mutiny Charocal</a></h3>
+            <h3 class="lot__title"><a class="text-link" href="lot.php">Куртка для сноуборда DC Mutiny Charocal</a></h3>
             <div class="lot__state">
               <div class="lot__rate">
                 <span class="lot__amount">12 ставок</span>
@@ -156,7 +156,7 @@
           </div>
           <div class="lot__info">
             <span class="lot__category">Разное</span>
-            <h3 class="lot__title"><a class="text-link" href="lot.html">Маска Oakley Canopy</a></h3>
+            <h3 class="lot__title"><a class="text-link" href="lot.php">Маска Oakley Canopy</a></h3>
             <div class="lot__state">
               <div class="lot__rate">
                 <span class="lot__amount">Стартовая цена</span>
