+++ pages/my-bets.html
@@ -62,7 +62,7 @@
             <div class="rates__img">
               <img src="../img/rate1.jpg" width="54" height="40" alt="Сноуборд">
             </div>
-            <h3 class="rates__title"><a href="lot.html">2014 Rossignol District Snowboard</a></h3>
+            <h3 class="rates__title"><a href="lot.php">2014 Rossignol District Snowboard</a></h3>
           </td>
           <td class="rates__category">
             Доски и лыжи
@@ -82,7 +82,7 @@
             <div class="rates__img">
               <img src="../img/rate2.jpg" width="54" height="40" alt="Сноуборд">
             </div>
-            <h3 class="rates__title"><a href="lot.html">DC Ply Mens 2016/2017 Snowboard</a></h3>
+            <h3 class="rates__title"><a href="lot.php">DC Ply Mens 2016/2017 Snowboard</a></h3>
           </td>
           <td class="rates__category">
             Доски и лыжи
@@ -103,7 +103,7 @@
               <img src="../img/rate3.jpg" width="54" height="40" alt="Крепления">
             </div>
             <div>
-              <h3 class="rates__title"><a href="lot.html">Крепления Union Contact Pro 2015 года размер L/XL</a></h3>
+              <h3 class="rates__title"><a href="lot.php">Крепления Union Contact Pro 2015 года размер L/XL</a></h3>
               <p>Телефон +7 900 667-84-48, Скайп: Vlas92. Звонить с 14 до 20</p>
             </div>
           </td>
@@ -125,7 +125,7 @@
             <div class="rates__img">
               <img src="../img/rate4.jpg" width="54" height="40" alt="Ботинки">
             </div>
-            <h3 class="rates__title"><a href="lot.html">Ботинки для сноуборда DC Mutiny Charocal</a></h3>
+            <h3 class="rates__title"><a href="lot.php">Ботинки для сноуборда DC Mutiny Charocal</a></h3>
           </td>
           <td class="rates__category">
             Ботинки
@@ -145,7 +145,7 @@
             <div class="rates__img">
               <img src="../img/rate5.jpg" width="54" height="40" alt="Куртка">
             </div>
-            <h3 class="rates__title"><a href="lot.html">Куртка для сноуборда DC Mutiny Charocal</a></h3>
+            <h3 class="rates__title"><a href="lot.php">Куртка для сноуборда DC Mutiny Charocal</a></h3>
           </td>
           <td class="rates__category">
             Одежда
@@ -165,7 +165,7 @@
             <div class="rates__img">
               <img src="../img/rate6.jpg" width="54" height="40" alt="Маска">
             </div>
-            <h3 class="rates__title"><a href="lot.html">Маска Oakley Canopy</a></h3>
+            <h3 class="rates__title"><a href="lot.php">Маска Oakley Canopy</a></h3>
           </td>
           <td class="rates__category">
             Разное
@@ -185,7 +185,7 @@
             <div class="rates__img">
               <img src="../img/rate7.jpg" width="54" height="40" alt="Сноуборд">
             </div>
-            <h3 class="rates__title"><a href="lot.html">DC Ply Mens 2016/2017 Snowboard</a></h3>
+            <h3 class="rates__title"><a href="lot.php">DC Ply Mens 2016/2017 Snowboard</a></h3>
           </td>
           <td class="rates__category">
             Доски и лыжи
